Refactor tests into a single file
