+++ bagelbids/tests/test_cli_pheno.py
@@ -1,7 +1,10 @@
+import json
+
+import pandas as pd
 import pytest
 from typer.testing import CliRunner
 
-from bagelbids.cli import bagel
+from bagelbids.cli import bagel, are_inputs_compatible
 
 
 @pytest.fixture
@@ -40,3 +43,22 @@ def test_valid_but_non_neurobagel_dictionary_fails(runner, test_data, tmp_path):
                                catch_exceptions=False)
 
     assert "data dictionary is not a valid Neurobagel data dictionary" in str(val_err.value)
+
+
+@pytest.mark.parametrize("example,is_valid", [
+    ("example1", True),
+    ("example2", True),
+    ("example3", True),
+    ("example4", True),
+    ("example5", True),
+    ("example6", True),
+    ("example7", False),
+])
+def test_validate_input(test_data, example, is_valid):
+    """Assures that two individually valid input files also make sense together"""
+    pheno = pd.read_csv(test_data / f"{example}.tsv", sep="\t")
+    with open(test_data / f"{example}.json", "r") as f:
+        data_dict = json.load(f)
+
+    result = are_inputs_compatible(data_dict=data_dict, pheno_df=pheno)
+    assert result == is_valid
